Overhaul packaging, launching services within a `omicron1` branded zones (#686)
This PR provides support for launching and managing services as zones.

### Nexus
- Exposes types in a way which can interoperate with Sled Agent.

### Sled Agent
- Parse a config file, rather than taking arguments over CLI
- Add a "Service Manager", responsible for zones that run independently of datasets (e.g., Nexus, Oximeter)

#### Networking
- Allow specifying static IP addresses for Zone addresses
- In doing so: Add some utilities to `sled-agent/src/illumos/zone.rs` to ensure that link-local addresses are available for IPv6 addressing. This lets us provide static IPv6 addresses for services, which we use elsewhere in the PR when setting up zones.

#### Storage
- Adds a "filesystem_put" endpoint to the Sled Agent, which allows a caller to request the creation of a local filesystem.
- Adds a "service_put" endpoint to the Sled Agent, which allows a caller to request the creation of a local service.
- Enable installation of `omicron1` branded zones.
- Overhaul the process of auto-loading installed zones and managing running zones.

#### Bootstrapping
- Enable the parsing of a `config-rss.toml` file, which provides requests that would *normally* arrive from the Rack Setup Service.
- Process the "rss requests" in the bootstrap agent, where they are injected before starting any local services.

### Packaging Tools
- Rely on [omicron-package](https://github.com/oxidecomputer/omicron-package) for package construction.
- Adds a new "zone" type package to omicron-package, allowing the creation of `omicron1`-branded `*.tar.gz` files.
- Adds a "cockroachdb", "clickhouse" packages to `package-manifest.toml`, with an associated `manifest.xml` file.

+++ package/src/bin/omicron-package.rs
@@ -6,28 +6,30 @@
  * Utility for bundling target binaries as tarfiles.
  */
 
-use omicron_common::packaging::sha256_digest;
-
 use anyhow::{anyhow, bail, Context, Result};
 use omicron_package::{parse, SubCommand};
+use omicron_zone_package::package::Package;
 use rayon::prelude::*;
-use reqwest;
 use serde_derive::Deserialize;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 use std::env;
-use std::fs::{create_dir_all, OpenOptions};
+use std::fs::create_dir_all;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use structopt::StructOpt;
-use tar::Builder;
-use tokio::io::AsyncWriteExt;
 
-const S3_BUCKET: &str = "https://oxide-omicron-build.s3.amazonaws.com";
+/// Describes the configuration for a set of packages.
+#[derive(Deserialize, Debug)]
+pub struct Config {
+    /// Packages to be built and installed.
+    #[serde(default, rename = "package")]
+    pub packages: BTreeMap<String, Package>,
 
-// Name for the directory component where additional packaged files are stored.
-const PKG: &str = "pkg";
-// Name for the directory component where downloaded blobs are stored.
-const BLOB: &str = "blob";
+    /// Packages to be installed, but which have been created outside this
+    /// repository.
+    #[serde(default, rename = "external_package")]
+    pub external_packages: BTreeMap<String, Package>,
+}
 
 #[derive(Debug, StructOpt)]
 #[structopt(name = "packaging tool")]
@@ -68,214 +70,48 @@ fn run_cargo_on_package(
     Ok(())
 }
 
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "lowercase")]
-enum Build {
-    /// The package is a rust binary, which should be compiled and extracted.
-    Rust,
-}
-
-#[derive(Deserialize, Debug)]
-struct PackageInfo {
-    // The name of the compiled binary to be used.
-    // TODO: Could be extrapolated to "produced build artifacts", we don't
-    // really care about the individual binary file.
-    binary_name: String,
-    // The name of the service name to be used on the target OS.
-    // Also used as a lookup within the "smf/" directory.
-    service_name: String,
-    // Instructs how this package should be compiled.
-    build: Build,
-    // A list of blobs from the Omicron build S3 bucket which should be placed
-    // within this package.
-    blobs: Option<Vec<PathBuf>>,
-    // If present, this service is necessary for bootstrapping, and should
-    // be durably installed to the target system.
-    //
-    // This is typically set to None - most services are bootstrapped
-    // by the bootstrap agent.
-    bootstrap: Option<PathBuf>,
-}
-
-impl PackageInfo {
-    fn binary_name(&self) -> &str {
-        &self.binary_name
-    }
-
-    // Returns the path to the compiled binary.
-    fn binary_path(&self, release: bool) -> PathBuf {
-        match &self.build {
-            Build::Rust => format!(
-                "target/{}/{}",
-                if release { "release" } else { "debug" },
-                self.binary_name()
-            )
-            .into(),
-        }
-    }
-
-    // Builds the requested package.
-    fn build(&self, package_name: &str, release: bool) -> Result<()> {
-        match &self.build {
-            Build::Rust => run_cargo_on_package("build", package_name, release),
-        }
-    }
-
-    // Checks the requested package.
-    fn check(&self, package_name: &str) -> Result<()> {
-        match &self.build {
-            Build::Rust => run_cargo_on_package("check", package_name, false),
-        }
-    }
-}
-
-#[derive(Deserialize, Debug)]
-struct Config {
-    // Path to SMF manifest directory.
-    smf: PathBuf,
-
-    #[serde(default, rename = "package")]
-    packages: BTreeMap<String, PackageInfo>,
-}
-
 async fn do_check(config: &Config) -> Result<()> {
     for (package_name, package) in &config.packages {
-        println!("Checking {}", package_name);
-        package.check(&package_name)?;
+        if let Some(rust_pkg) = &package.rust {
+            println!("Checking {}", package_name);
+            run_cargo_on_package("check", &package_name, rust_pkg.release)?;
+        }
     }
     Ok(())
 }
 
-async fn do_package(
-    config: &Config,
-    output_directory: &Path,
-    release: bool,
-) -> Result<()> {
-    // Create the output directory, if it does not already exist.
+async fn do_package(config: &Config, output_directory: &Path) -> Result<()> {
     create_dir_all(&output_directory)
         .map_err(|err| anyhow!("Cannot create output directory: {}", err))?;
 
-    // As we emit each package bundle, capture their digests for
-    // verification purposes.
-    let mut digests = HashMap::<String, Vec<u8>>::new();
-
     for (package_name, package) in &config.packages {
-        println!("Building {}", package_name);
-        package.build(&package_name, release)?;
-
-        let tarfile =
-            output_directory.join(format!("{}.tar", package.service_name));
-        let file = OpenOptions::new()
-            .write(true)
-            .read(true)
-            .truncate(true)
-            .create(true)
-            .open(&tarfile)
-            .map_err(|err| anyhow!("Cannot create tarfile: {}", err))?;
-
-        // Create an archive filled with:
-        // - The binary
-        // - The corresponding SMF directory
-        //
-        // TODO: We could add compression here, if we'd like?
-        let mut archive = Builder::new(file);
-        archive.mode(tar::HeaderMode::Deterministic);
-
-        // Add binary
-        archive
-            .append_path_with_name(
-                package.binary_path(release),
-                &package.binary_name,
-            )
-            .map_err(|err| {
-                anyhow!("Cannot append binary to tarfile: {}", err)
-            })?;
-
-        // Add SMF directory
-        let smf_path: PathBuf = format!(
-            "{}/{}",
-            config.smf.to_string_lossy(),
-            package.service_name
-        )
-        .into();
-        add_path_to_archive(&mut archive, &smf_path, Path::new(PKG))?;
-
-        // Add (and possibly download) blobs
-        add_blobs(&mut archive, package_name, package, output_directory)
-            .await?;
-
-        let mut file = archive
-            .into_inner()
-            .map_err(|err| anyhow!("Failed to finalize archive: {}", err))?;
-
-        // Once we've created the archive, acquire a digest which can
-        // later be used for verification.
-        let digest = sha256_digest(&mut file)?;
-        digests.insert(package.binary_name().into(), digest.as_ref().into());
-    }
-
-    let toml = toml::to_string(&digests)?;
-    std::fs::write(output_directory.join("digest.toml"), &toml)?;
-    Ok(())
-}
-
-// Adds all files within "path" to "archive".
-//
-// Within the archive, all files are renamed to "dst_prefix/<file_name>".
-fn add_path_to_archive(
-    archive: &mut Builder<std::fs::File>,
-    path: &Path,
-    dst_prefix: &Path,
-) -> Result<()> {
-    for entry in walkdir::WalkDir::new(&path) {
-        let entry =
-            entry.map_err(|err| anyhow!("Cannot access entry: {}", err))?;
-        if entry.file_name().to_string_lossy().starts_with('.') {
-            // Omit hidden files - we don't want to include text editor
-            // artifacts in the tarfile.
-            continue;
+        if let Some(rust_pkg) = &package.rust {
+            println!("Building: {}", package_name);
+            run_cargo_on_package("build", package_name, rust_pkg.release)?;
         }
-        let dst = dst_prefix.join(entry.path().strip_prefix(&path)?);
-        println!(
-            "Archiving {} -> {}",
-            entry.path().to_string_lossy(),
-            dst.to_string_lossy()
-        );
-        archive
-            .append_path_with_name(entry.path(), dst)
-            .map_err(|err| anyhow!("Cannot append file to tarfile: {}", err))?;
+        package.create(&output_directory).await?;
     }
-    Ok(())
-}
 
-async fn add_blobs(
-    archive: &mut Builder<std::fs::File>,
-    name: &String,
-    package: &PackageInfo,
-    output_directory: &Path,
-) -> Result<()> {
-    if let Some(blobs) = &package.blobs {
-        let blobs_path = output_directory.join(name);
-        std::fs::create_dir_all(&blobs_path)?;
-        for blob in blobs {
-            let blob_path = blobs_path.join(blob);
-            // TODO: Check against hash, download if mismatch (i.e.,
-            // corruption/update).
-            if !blob_path.exists() {
-                download(&blob.to_string_lossy(), &blob_path).await?;
-            }
+    for (package_name, package) in &config.external_packages {
+        let path = package.get_output_path(&output_directory);
+        if !path.exists() {
+            bail!(
+                "The package for {} (expected at {}) does not exist.
+Omicron can build all packages that exist within the repository, but packages
+from outside the repository must be created manually, and copied to the output
+directory (which is currently: {}).
+
+This is currently a shortcoming with the build system - if you have an interest
+in improving the cross-repository meta-build system, please contact sean@.",
+                package_name,
+                path.to_string_lossy(),
+                std::fs::canonicalize(output_directory)
+                    .unwrap()
+                    .to_string_lossy(),
+            );
         }
-        add_path_to_archive(archive, blobs_path.as_path(), Path::new(BLOB))?;
     }
-    Ok(())
-}
 
-// Downloads "source" from S3_BUCKET to "destination".
-async fn download(source: &str, destination: &Path) -> Result<()> {
-    println!("Downloading {} to {}", source, destination.to_string_lossy());
-    let response = reqwest::get(format!("{}/{}", S3_BUCKET, source)).await?;
-    let mut file = tokio::fs::File::create(destination).await?;
-    file.write_all(&response.bytes().await?).await?;
     Ok(())
 }
 
@@ -288,27 +124,20 @@ fn do_install(
         anyhow!("Cannot create installation directory: {}", err)
     })?;
 
-    println!(
-        "Copying digest.toml from {} to {}",
-        artifact_dir.to_string_lossy(),
-        install_dir.to_string_lossy()
-    );
-    // Move the digest of expected packages.
-    std::fs::copy(
-        artifact_dir.join("digest.toml"),
-        install_dir.join("digest.toml"),
-    )?;
-
     // Copy all packages to the install location in parallel.
-    let packages: Vec<(&String, &PackageInfo)> =
-        config.packages.iter().collect();
+    let packages: Vec<(&String, &Package)> =
+        config.packages.iter().chain(config.external_packages.iter()).collect();
     packages.into_par_iter().try_for_each(|(_, package)| -> Result<()> {
-        let tarfile =
-            artifact_dir.join(format!("{}.tar", package.service_name));
+        let tarfile = if package.zone {
+            artifact_dir.join(format!("{}.tar.gz", package.service_name))
+        } else {
+            artifact_dir.join(format!("{}.tar", package.service_name))
+        };
+
         let src = tarfile.as_path();
         let dst = install_dir.join(src.strip_prefix(artifact_dir)?);
         println!(
-            "Installing {} -> {}",
+            "Installing Service: {} -> {}",
             src.to_string_lossy(),
             dst.to_string_lossy()
         );
@@ -321,33 +150,31 @@ fn do_install(
 
     // Extract and install the bootstrap service, which itself extracts and
     // installs other services.
-    for (_, package) in &config.packages {
-        if let Some(manifest) = &package.bootstrap {
-            let tar_path =
-                install_dir.join(format!("{}.tar", package.service_name));
-            let service_path = install_dir.join(&package.service_name);
-            println!(
-                "Unpacking {} to {}",
-                tar_path.to_string_lossy(),
-                service_path.to_string_lossy()
-            );
+    if let Some(package) = config.packages.get("omicron-sled-agent") {
+        let tar_path =
+            install_dir.join(format!("{}.tar", package.service_name));
+        let service_path = install_dir.join(&package.service_name);
+        println!(
+            "Unpacking {} to {}",
+            tar_path.to_string_lossy(),
+            service_path.to_string_lossy()
+        );
 
-            let tar_file = std::fs::File::open(&tar_path)?;
-            let _ = std::fs::remove_dir_all(&service_path);
-            std::fs::create_dir_all(&service_path)?;
-            let mut archive = tar::Archive::new(tar_file);
-            archive.unpack(&service_path)?;
+        let tar_file = std::fs::File::open(&tar_path)?;
+        let _ = std::fs::remove_dir_all(&service_path);
+        std::fs::create_dir_all(&service_path)?;
+        let mut archive = tar::Archive::new(tar_file);
+        archive.unpack(&service_path)?;
 
-            let mut manifest_path = install_dir.to_path_buf();
-            manifest_path.push(&package.service_name);
-            manifest_path.push(PKG);
-            manifest_path.push(manifest);
-            println!(
-                "Installing bootstrap service from {}",
-                manifest_path.to_string_lossy()
-            );
-            smf::Config::import().run(&manifest_path)?;
-        }
+        let manifest_path = install_dir
+            .join(&package.service_name)
+            .join("pkg")
+            .join("manifest.xml");
+        println!(
+            "Installing bootstrap service from {}",
+            manifest_path.to_string_lossy()
+        );
+        smf::Config::import().run(&manifest_path)?;
     }
 
     Ok(())
@@ -355,12 +182,20 @@ fn do_install(
 
 // Attempts to both disable and delete all requested packages.
 fn uninstall_all_packages(config: &Config) {
-    for package in config.packages.values() {
-        let _ = smf::Adm::new()
-            .disable()
-            .synchronous()
-            .run(smf::AdmSelection::ByPattern(&[&package.service_name]));
-        let _ = smf::Config::delete().force().run(&package.service_name);
+    for package in
+        config.packages.values().chain(config.external_packages.values())
+    {
+        if package.zone {
+            // TODO(https://github.com/oxidecomputer/omicron/issues/723):
+            // At the moment, zones are entirely managed by the sled agent,
+            // but could be removed here.
+        } else {
+            let _ = smf::Adm::new()
+                .disable()
+                .synchronous()
+                .run(smf::AdmSelection::ByPattern(&[&package.service_name]));
+            let _ = smf::Config::delete().force().run(&package.service_name);
+        }
     }
 }
 
@@ -401,8 +236,8 @@ async fn main() -> Result<()> {
     }
 
     match &args.subcommand {
-        SubCommand::Package { artifact_dir, release } => {
-            do_package(&config, &artifact_dir, *release).await?;
+        SubCommand::Package { artifact_dir } => {
+            do_package(&config, &artifact_dir).await?;
         }
         SubCommand::Check => do_check(&config).await?,
         SubCommand::Install { artifact_dir, install_dir } => {
